Refactor Chunk reloading so that Forge doesn't need to rebake all models
