+++ common/src/main/resources/fiblib-common.mixins.json
@@ -1,0 +1,14 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "dev.hephaestus.fiblib.mixin",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": [
+    "ChunkReloader"
+  ],
+  "client": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
