+++ forge/src/main/java/dev/hephaestus/fiblib/forge/PlatformImpl.java
@@ -1,13 +1,10 @@
 package dev.hephaestus.fiblib.forge;
 
-import dev.hephaestus.fiblib.FibLib;
 import dev.hephaestus.fiblib.Platform;
-import dev.hephaestus.fiblib.forge.network.ServerPacketHandler;
 import net.minecraft.server.level.ServerPlayer;
 
 public class PlatformImpl extends Platform {
     public static void resendChunks(ServerPlayer player) {
-        FibLib.log("Trigger reload chunk");
-        ServerPacketHandler.getInstance().sendReloadPacket(player);
+        resendChunksVanilla(player);
     }
 }
